+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Requires `innmind/foundation:~1.0`
 - `Innmind\S3\Bucket::upload()` now returns an `Innmind\Immutable\Attempt<Innmind\Immutable\SideEffect>`
 - `Innmind\S3\Bucket::delete()` now returns an `Innmind\Immutable\Attempt<Innmind\Immutable\SideEffect>`
+- `Innmind\S3\Format\*` classes have grouped into the enum `Innmind\S3\Format\Amazon`
 
 ### Fixed